feat: write-tree 명령어 옵션이 들어온 경우 에러 처리 및 함수 분리
- writeTreeUtil.kt 추가 및 writeTree 함수 분리
